create and delete monitor's dce

+++ azure/resources-destroy.ps1
@@ -1,6 +1,18 @@
 $scriptRoot = $PSScriptRoot
 . $scriptRoot\resources-lib-commands.ps1
 
+function Test-User-Acceptance([Parameter(Mandatory = $true)][string]$message) {
+    Write-Host $message -foregroundcolor Yellow
+    Write-Host "(Y to accept otherwise any other caracter)" -foregroundcolor Gray
+    $key = $Host.UI.RawUI.ReadKey()
+
+    $accepted = "$($key.Character)".ToLower() -eq "y"
+    if ( $accepted) { Write-Host " -> accepted !" -ForegroundColor Yellow }
+    else { Write-Host " -> refused !" -ForegroundColor Yellow }
+    
+    return $accepted
+}
+
 try {
     Login | Out-Null
 
@@ -8,31 +20,44 @@ try {
     $settings = Get-Settings
     
     $settings.subscription | Set-Subscription
-    Write-Host "subscription : `n$($settings.subscription | Get-Subscription | ConvertTo-Json)"
+    $subscription = $settings.subscription | Get-Subscription
+    Write-Host "subscription : `n$($subscription | ConvertTo-Json | Format-Json)"
 
-    $resource = $settings.resources | Where-Object { $_.kind -eq "resource group" }
+    # ACT : delete Monitor resources : dce 
+    if (Test-User-Acceptance "Do you want to proceed with deleting the monitor endpoints?") {
+        $settings.resources `
+        | Where-Object { $_.kind -eq "monitor data-collection endpoint" } `
+        | Foreach-Object {
+            Write-Host "destroy monitor's data collection $($_.name)" -ForegroundColor Yellow
+            az monitor data-collection endpoint delete --yes --name $_.name --resource-group $_.resourceGroup --subscription $subscription.id
+        }
+    }
     
-    # ACT : delete service principals
-    $list = $settings.resources | where-object { $_.servicePrincipal -ne $null } | Select-Object -ExpandProperty servicePrincipal
-    foreach ($item in $list) {
-        $servicePrincipal = $item | Get-ServicePrincipal
-        if ($null -eq $servicePrincipal) { continue }
-        
-        Write-Host "destroy service principal $($servicePrincipal.appDisplayName) (id: '$($servicePrincipal.appId)')" -ForegroundColor Yellow
-        # quelque chose est étrange !!    cela ne fonctionne pas !! =>         
-        az ad sp delete --id $servicePrincipal.appId
+    # ACT : delete resources    
+    if (Test-User-Acceptance "Do you want to proceed with deleting the resources?") {
+        $resource = $settings.resources | Where-Object { $_.kind -eq "resource group" }
+        if ($null -eq ($resource | Get-RessourceGroup)) {
+            Write-Host "resource group not found, nothing to delete" -ForegroundColor Yellow
+            return
+        }
+        Write-Host "destroy resource group $($resource.name)" -ForegroundColor Yellow
+        az group delete -n $resource.name --yes
     }
-
-    # ACT : delete resources
-    if ($null -eq ($resource | Get-RessourceGroup)) {
-        Write-Host "resource group not found, nothing to delete" -ForegroundColor Yellow
-        return
+    
+    # ACT : delete service principals    
+    if (Test-User-Acceptance "Do you want to proceed with deleting the service principals?") {
+        $settings.resources `
+        | Where-Object { $_.servicePrincipal -ne $null } `
+        | ForEach-Object {
+            $sp = $_.servicePrincipal | Get-ServicePrincipal
+            if ($sp -eq $null) {
+                Write-Host "service principal not found, nothing to delete" -ForegroundColor Yellow
+                return
+            }
+            Write-Host "destroy service principal $($sp.appDisplayName) (id: '$($sp.appId)')" -ForegroundColor Yellow
+            az ad sp delete --id $sp.appId
+        } 
     }
-
-    Write-Host "destroy resource group $($resource.name)" -ForegroundColor Yellow
-    az group delete -n $resource.name --yes
-
-    Write-Host "resources destroyed" -foregroundcolor Green
 }
 catch {
     Write-Error $Error[0]
